requirements.txt
# Existing dependencies
appdirs==1.4.4
argcomplete==3.6.3
importlib_resources==6.5.2
pillow
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
setuptools==80.9.0
six==1.17.0

# Web frameworks
fastapi==0.115.5
uvicorn[standard]==0.32.1
streamlit==1.40.2

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Email
sendgrid==6.11.0

# Text processing
emoji==2.14.0
unidecode==1.3.8

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic[email]
